docker-environments/docker-basics/apps/python-app/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
redis==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
Werkzeug==2.3.7